Move pluralize function to helpers
