+++ src/commands/sourcemaps/renderer.ts
@@ -2,7 +2,7 @@ import chalk from 'chalk'
 
 import {ICONS} from '../../helpers/formatting'
 import {UploadStatus} from '../../helpers/upload'
-import {pluralize} from './common/utils'
+import {pluralize} from '../../helpers/utils'
 import {Sourcemap} from './interfaces'
 
 export const renderGitWarning = (errorMessage: string) =>
